Move url filter into its own file and switching to pleaseDrop _sampling property that gets propogated through filtered express/http server requests
